mcp-servers/photo-ranker/pyproject.toml
[project]
name = "photo-ranker"
version = "0.1.0"
description = "Photo ranking and classification MCP server for OpenClaw"
requires-python = ">=3.12"
dependencies = [
    "mcp>=1.0.0",
    "pillow>=10.0",
    "numpy>=1.26",
    "pydantic>=2.0",
    "imagehash>=4.3",
]

[project.optional-dependencies]
vlm = ["mlx-vlm>=0.1"]
aesthetic = ["open-clip-torch>=2.24", "torch>=2.0"]
face = ["mediapipe>=0.10"]
face-legacy = ["face-recognition>=1.3"]
all = [
    "photo-ranker[vlm,aesthetic,face]",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
